[FIX] #183 네비게이션 및 safe args 수정 (nullable 항목 없애기)
