+++ app/src/main/res/navigation/nav_graph.xml
@@ -39,9 +39,6 @@
         android:label="fragment_my_page"
         tools:layout="@layout/fragment_my_page">
         <action
-            android:id="@+id/action_myPageFragment_to_userProfileEditFragment"
-            app:destination="@id/userProfileEditFragment" />
-        <action
             android:id="@+id/action_myPageFragment_to_settingsFragment"
             app:destination="@id/settingsFragment" />
         <action
@@ -56,24 +53,9 @@
         <action
             android:id="@+id/action_myPageFragment_to_inquiryFragment"
             app:destination="@id/inquiryFragment" />
-    </fragment>
-    <fragment
-        android:id="@+id/userProfileEditFragment"
-        android:name="com.starters.hsge.presentation.main.mypage.UserProfileEditFragment"
-        android:label="fragment_user_profile_edit"
-        tools:layout="@layout/fragment_user_profile_edit" >
-        <argument
-            android:name="userProfileInfo"
-            app:argType="com.starters.hsge.domain.model.UserProfileInfo"
-            app:nullable="true" />
         <action
-            android:id="@+id/action_userProfileEditFragment_to_userProfileEditIconFragment"
-            app:destination="@id/userProfileEditIconFragment" />
-        <action
-            android:id="@+id/action_userProfileEditFragment_to_myPageFragment"
-            app:destination="@id/myPageFragment"
-            app:popUpTo="@id/myPageFragment"
-            app:popUpToInclusive="true" />
+            android:id="@+id/action_myPageFragment_to_userProfileEditFragment"
+            app:destination="@id/userProfileEditFragment" />
     </fragment>
     <fragment
         android:id="@+id/settingsFragment"
@@ -137,21 +119,6 @@
              />
     </fragment>
     <fragment
-        android:id="@+id/userProfileEditIconFragment"
-        android:name="com.starters.hsge.presentation.register.fragment.UserProfileIconFragment"
-        android:label="UserProfileEditIconFragment"
-        tools:layout="@layout/fragment_user_profile_icon">
-        <action
-            android:id="@+id/action_userProfileEditIconFragment_to_userProfileEditFragment"
-            app:destination="@id/userProfileEditFragment"
-            app:popUpTo="@id/userProfileEditFragment"
-            app:popUpToInclusive="true" />
-        <argument
-            android:name="checkLayout"
-            app:argType="integer"
-            android:defaultValue="0" />
-    </fragment>
-    <fragment
         android:id="@+id/addDogProfileFragment"
         android:name="com.starters.hsge.presentation.main.add.AddDogProfileFragment"
         android:label="fragment_add_dog_profile"
@@ -169,7 +136,7 @@
             app:destination="@id/partnerDogsFragment" />
         <argument
             android:name="chatInfo"
-            app:argType="com.starters.hsge.data.model.remote.response.ChatListResponse" />
+            app:argType="com.starters.hsge.domain.model.ChatListInfo" />
     </fragment>
     <fragment
         android:id="@+id/chatReportFragment"
@@ -209,4 +176,32 @@
         android:name="com.starters.hsge.presentation.main.mypage.InquiryFragment"
         android:label="InquiryFragment"
         tools:layout="@layout/fragment_inquiry"/>
+    <fragment
+        android:id="@+id/userProfileEditFragment"
+        android:name="com.starters.hsge.presentation.main.user.UserProfileEditFragment"
+        android:label="UserProfileEditFragment" >
+        <action
+            android:id="@+id/action_userProfileEditFragment_to_profileIconFragment2"
+            app:destination="@id/profileIconFragment2" />
+        <argument
+            android:name="userProfileInfo"
+            app:argType="com.starters.hsge.domain.model.UserProfileInfo"
+            app:nullable="true" />
+    </fragment>
+    <fragment
+        android:id="@+id/profileIconFragment2"
+        android:name="com.starters.hsge.presentation.register.fragment.ProfileIconFragment"
+        android:label="ProfileIconFragment" >
+        <argument
+            android:name="checkLayout"
+            app:argType="integer" />
+        <action
+            android:id="@+id/action_profileIconFragment2_to_userProfileEditFragment"
+            app:destination="@id/userProfileEditFragment"
+            app:popUpTo="@id/userProfileEditFragment"
+            app:popUpToInclusive="true" />
+        <argument
+            android:name="nickName"
+            app:argType="string" />
+    </fragment>
 </navigation>
