+++ app/src/main/res/navigation/nav_graph_register.xml
@@ -29,8 +29,8 @@
             android:id="@+id/action_userImageFragment_to_dogPhotoFragment"
             app:destination="@id/dogPhotoFragment" />
         <action
-            android:id="@+id/action_userImageFragment_to_userProfileIconFragment"
-            app:destination="@id/userProfileIconFragment" />
+            android:id="@+id/action_userImageFragment_to_profileIconFragment"
+            app:destination="@id/profileIconFragment" />
     </fragment>
     <fragment
         android:id="@+id/userNickNameFragment"
@@ -40,21 +40,6 @@
         <action
             android:id="@+id/action_userNickNameFragment_to_userImageFragment"
             app:destination="@id/userImageFragment" />
-    </fragment>
-    <fragment
-        android:id="@+id/userProfileIconFragment"
-        android:name="com.starters.hsge.presentation.register.fragment.UserProfileIconFragment"
-        android:label="UserProfileIconFragment"
-        tools:layout="@layout/fragment_user_profile_icon">
-        <action
-            android:id="@+id/action_userProfileIconFragment_to_userImageFragment"
-            app:destination="@id/userImageFragment"
-            app:popUpTo="@id/userImageFragment"
-            app:popUpToInclusive="true"/>
-        <argument
-            android:name="checkLayout"
-            app:argType="integer"
-            android:defaultValue="0" />
     </fragment>
     <fragment
         android:id="@+id/dogNameFragment"
@@ -110,5 +95,12 @@
             android:id="@+id/action_dogDislikeTagFragment_to_userLocationFragment"
             app:destination="@id/userLocationFragment" />
     </fragment>
-
+    <fragment
+        android:id="@+id/profileIconFragment"
+        android:name="com.starters.hsge.presentation.register.fragment.ProfileIconFragment"
+        android:label="ProfileIconFragment" >
+        <argument
+            android:name="checkLayout"
+            app:argType="integer" />
+    </fragment>
 </navigation>
